Add expire for Memcache and Memcached lock storage
