+++ src/NinjaMutex/Lock/MemcacheLock.php
@@ -16,13 +16,84 @@ use Memcache;
  *
  * @author Kamil Dziedzic <arvenil@klecza.pl>
  */
-class MemcacheLock extends MemcacheLockAbstract
+class MemcacheLock extends LockAbstract
 {
     /**
-     * @param Memcache $memcache
+     * Maximum expiration time in seconds (30 days)
+     * http://php.net/manual/en/memcache.add.php
      */
-    public function __construct(Memcache $memcache)
+    const MAX_EXPIRATION = 2592000;
+
+    /**
+     * Memcache connection
+     *
+     * @var Memcache
+     */
+    protected $memcache;
+
+    /**
+     * @var int Expiration time of the lock in seconds
+     */
+    protected $expiration = 0;
+
+    /**
+     * @param Memcache $memcache
+     * @param int      $expiration Expiration time of the lock in seconds. If it's equal to zero (default), the lock will never expire.
+     *                             Max 2592000s (30 days), if greater it will be capped to 2592000 without throwing an error.
+     *                             WARNING: Using value higher than 0 may lead to race conditions. If you set too low expiration time
+     *                             e.g. 30s and critical section will run for 31s another process will gain lock at the same time,
+     *                             leading to unpredicted behaviour. Use with caution.
+     */
+    public function __construct(Memcache $memcache, $expiration = 0)
     {
-        parent::__construct($memcache);
+        parent::__construct();
+
+        $this->memcache = $memcache;
+        if ($expiration > static::MAX_EXPIRATION) {
+            $expiration = static::MAX_EXPIRATION;
+        }
+        $this->timeout = $expiration;
+    }
+
+    /**
+     * @param  string $name name of lock
+     * @param  bool   $blocking
+     * @return bool
+     */
+    protected function getLock($name, $blocking)
+    {
+        if (!$this->memcache->add($name, serialize($this->getLockInformation()), 0, $this->expiration)) {
+            return false;
+        }
+
+        return true;
+    }
+
+    /**
+     * Release lock
+     *
+     * @param  string $name name of lock
+     * @return bool
+     */
+    public function releaseLock($name)
+    {
+        if (isset($this->locks[$name]) && $this->memcache->delete($name)) {
+            unset($this->locks[$name]);
+
+            return true;
+        }
+
+        return false;
+    }
+
+    /**
+     * Check if lock is locked
+     *
+     * @param  string $name name of lock
+     * @return bool
+     */
+    public function isLocked($name)
+    {
+        return false !== $this->memcache->get($name);
     }
 }
